Add syntax table for comments
Some packages modify the font-lock of text by looking at the syntax
table of the major-mode attached to the buffer (i.e.
rainbow-delimiters). If we want the tree-sitter parser to properly
highlight things AND not have other packages modify the font-locking
once it is done, we need to set the syntax-table to reflect this.

If this step was not done, someone used tree-sitter AND used
rainbow-delimiters, then they would get proper syntax highlighting
inside a block comment, but the ( and ) of the block comment would be
highlighted as normal code/text, which is obviously incorrect.